requirements.txt
pytube>=12.1.2
google-api-python-client>=2.94.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
moviepy>=1.0.3
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.2
pandas>=2.0.3
lxml>=4.9.3
yt-dlp>=2023.11.16
psutil>=5.9.5
flask>=2.3.3
werkzeug>=2.3.7
flask-cors>=4.0.0
opencv-python>=4.8.1
pillow>=10.0.1
gunicorn>=21.2.0
aiohttp>=3.8.5
asyncio>=3.4.3
tqdm>=4.66.1
colorama>=0.4.6
numpy>=1.24.4
matplotlib>=3.7.2